Fix long int overflow on some platforms
Unfortunately many platforms define the datatype long as 32bit integer
value. This lead to integer overflows especially on windows. I am using
long long instead of long now. Time to move to rust ;)

Using ini file format for config file on windows

+++ src/labpowermodel.cpp
@@ -107,7 +107,7 @@ void LabPowerModel::setOCP(bool status) { this->status->setOcp(status); }
 bool LabPowerModel::getOCP() { return this->status->getOcp(); }
 void LabPowerModel::setOTP(bool status) { this->status->setOtp(status); }
 bool LabPowerModel::getOTP() { return this->status->getOtp(); }
-long LabPowerModel::getDuration() { return this->status->getDuration(); }
+long long LabPowerModel::getDuration() { return this->status->getDuration(); }
 std::vector<std::shared_ptr<PowerSupplyStatus>> LabPowerModel::getBuffer()
 {
     return this->statusBuffer;
